+++ AirSeaFluxCode/hum_subs/__init__.py
@@ -1,0 +1,6 @@
+from .vapor_pressure import VaporPressure
+from .qsat import qsat_air, qsat_sea
+from .humidity import gamma, get_hum
+
+
+__all__ = ['gamma', 'get_hum', 'qsat_air', 'qsat_sea', 'VaporPressure']
