+++ AirSeaFluxCode/util_subs/__init__.py
@@ -1,0 +1,6 @@
+from .utils import (CtoK, kappa, get_heights, gc, visc_air,
+                    set_flag, get_outvars, rho_air)
+
+
+__all__ = ['CtoK', 'kappa', 'get_heights', 'gc', 'visc_air', 'set_flag',
+           'get_outvars', 'rho_air']
